+++ core/class/Telecommande/Connectors/ConnectorImpl_Legacy.php
@@ -4,18 +4,18 @@ if (defined('WEB')) {
     require_once __DIR__ . '/IConnector.php';
     
     require_once __DIR__ . '/../../Logger.class.php';
-    require_once __DIR__ . '/../TvDomSamsung.TvParametres.class.php';
-    require_once __DIR__ . '/../TvDomSamsung.CommunicationStatus.class.php';
-    require_once __DIR__ . '/../Sockets/TvDomSamsung.SocketWrapper.class.php';
-    require_once __DIR__ . '/LegacyTv/TvDomSamsung.MessageBuilder.class.php';
+    require_once __DIR__ . '/../JeeTizen.TvParametres.class.php';
+    require_once __DIR__ . '/../JeeTizen.CommunicationStatus.class.php';
+    require_once __DIR__ . '/../Sockets/JeeTizen.SocketWrapper.class.php';
+    require_once __DIR__ . '/LegacyTv/JeeTizen.MessageBuilder.class.php';
     
 } else {
     require_once __DIR__ . '/IConnector.php';
     require_once __DIR__ . '/../../Logger.class.php';
-    require_once __DIR__ . '/../TvDomSamsung.TvParametres.class.php';
-    require_once __DIR__ . '/../TvDomSamsung.CommunicationStatus.class.php';
-    require_once __DIR__ . '/../Sockets/TvDomSamsung.SocketWrapper.class.php';
-    require_once __DIR__ . '/LegacyTv/TvDomSamsung.MessageBuilder.class.php';
+    require_once __DIR__ . '/../JeeTizen.TvParametres.class.php';
+    require_once __DIR__ . '/../JeeTizen.CommunicationStatus.class.php';
+    require_once __DIR__ . '/../Sockets/JeeTizen.SocketWrapper.class.php';
+    require_once __DIR__ . '/LegacyTv/JeeTizen.MessageBuilder.class.php';
 }